refactor: add workers and main modules
