+++ python-api/python_api/config.py
@@ -1,13 +1,8 @@
 import os
-import sys
-import logging
-from typing import Optional, Any, Dict
-from contextvars import ContextVar
+from typing import Optional
 
 from dotenv import load_dotenv
-import structlog
 from fastapi import Header, HTTPException
-from prometheus_client import Histogram, Counter, Gauge, REGISTRY
 
 load_dotenv()
 
@@ -30,34 +25,6 @@ STUCK_ORDER_HOURS = int(os.getenv("STUCK_ORDER_HOURS", "24"))
 STUCK_CHECK_INTERVAL = int(os.getenv("STUCK_CHECK_INTERVAL", "600"))
 SIGNING_DEADLINE_DAYS = int(os.getenv("SIGNING_DEADLINE_DAYS", "7"))
 RATE_LIMIT = os.getenv("RATE_LIMIT", "100/minute")
-
-# ---- Prometheus metrics ----
-def _metric(name, factory):
-    existing = REGISTRY._names_to_collectors.get(name)
-    return existing if existing else factory()
-
-RPC_HIST = _metric('rpc_duration_seconds', lambda: Histogram('rpc_duration_seconds', 'Bitcoin Core RPC duration', ['method']))
-WEBHOOK_COUNTER = _metric('webhook_total', lambda: Counter('webhook_total', 'Webhook deliveries', ['status']))
-PENDING_SIG = _metric('pending_signatures', lambda: Gauge('pending_signatures', 'Open PSBT signatures'))
-STUCK_COUNTER = _metric('stuck_orders_total', lambda: Counter('stuck_orders_total', 'Orders stuck beyond threshold', ['state']))
-BROADCAST_FAIL = _metric('broadcast_fail_total', lambda: Counter('broadcast_fail_total', 'Failed transaction broadcasts'))
-WEBHOOK_QUEUE_SIZE = _metric('webhook_queue_size', lambda: Gauge('webhook_queue_size', 'Pending webhooks in queue'))
-
-# ---- logging ----
-logging.basicConfig(stream=sys.stdout, format="%(message)s", level=logging.INFO)
-structlog.configure(
-    processors=[
-        structlog.processors.TimeStamper(fmt="iso"),
-        structlog.processors.add_log_level,
-        structlog.processors.JSONRenderer(),
-    ],
-    logger_factory=structlog.stdlib.LoggerFactory(),
-)
-log = structlog.get_logger()
-
-req_id_var: ContextVar[Optional[str]] = ContextVar("request_id", default=None)
-order_id_var: ContextVar[Optional[str]] = ContextVar("order_id", default=None)
-actor_var: ContextVar[Optional[str]] = ContextVar("actor", default=None)
 
 # ---- State machine ----
 STATES = [
